Adding multiple playpen pages
#284
