+++ waltz-ng/client/playpen/1/playpen1.html
@@ -1,7 +1,7 @@
 <div>
 
     <waltz-page-header name="Playpen"
-                       small=""
+                       small="1"
                        icon="rocket">
         <ol class="waltz-breadcrumbs">
             <li><a ui-sref="main">Home</a></li>
@@ -9,13 +9,7 @@
         </ol>
     </waltz-page-header>
 
+    <waltz-section name="Section">
 
-    <!-- INDICATORS -->
-    <waltz-entity-statistic-summary-section name="Indicators"
-                                    entity-statistics="ctrl.entityStatistics">
-
-
-
-    </waltz-entity-statistic-summary-section>
-
+    </waltz-section>
 </div>
