+++ waltz-ng/client/playpen/2/playpen2.html
@@ -1,0 +1,15 @@
+<div>
+
+    <waltz-page-header name="Playpen"
+                       small="2"
+                       icon="rocket">
+        <ol class="waltz-breadcrumbs">
+            <li><a ui-sref="main">Home</a></li>
+            <li>Playpen</li>
+        </ol>
+    </waltz-page-header>
+
+    <waltz-section name="Section">
+
+    </waltz-section>
+</div>
