+++ waltz-ng/client/playpen/3/playpen3.html
@@ -1,0 +1,15 @@
+<div>
+
+    <waltz-page-header name="Playpen"
+                       small="3"
+                       icon="rocket">
+        <ol class="waltz-breadcrumbs">
+            <li><a ui-sref="main">Home</a></li>
+            <li>Playpen</li>
+        </ol>
+    </waltz-page-header>
+
+    <waltz-section name="Section">
+
+    </waltz-section>
+</div>
