+++ waltz-ng/client/playpen/list.html
@@ -1,0 +1,20 @@
+<div>
+
+    <waltz-page-header name="Playpen"
+                       small="List"
+                       icon="rocket">
+        <ol class="waltz-breadcrumbs">
+            <li><a ui-sref="main">Home</a></li>
+            <li>Playpen</li>
+        </ol>
+    </waltz-page-header>
+
+    <waltz-section name="Playpens">
+        <ul>
+            <li><a ui-sref="main.playpen.1">Playpen 1</a></li>
+            <li><a ui-sref="main.playpen.2">Playpen 2</a></li>
+            <li><a ui-sref="main.playpen.3">Playpen 3</a></li>
+        </ul>
+    </waltz-section>
+
+</div>
